requirements.txt

aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
redis==3.5.3
bcrypt==3.2.0
flask_sqlalchemy==2.4.4
flask_bcrypt==0.7.1
marshmallow==3.8.0
jwt==1.0.0
python-dotenv==0.14.0
psycopg2==2.8.6
flask-jwt-extended==3.24.1
